* arch re-arrange * cleanup
