+++ app/src/main/java/com/netchar/nicknamer/domen/NicknameModelsProvider.kt
@@ -1,0 +1,7 @@
+package com.netchar.nicknamer.domen
+
+import com.netchar.nicknamer.domen.models.NicknameModel
+
+interface NicknameModelsProvider {
+    fun getModels() : Map<String, NicknameModel>
+}
