+++ app/src/main/java/com/netchar/nicknamer/domen/service/NicknameGeneratorService.kt
@@ -16,14 +16,10 @@
 
 package com.netchar.nicknamer.domen.service
 
+import com.netchar.nicknamer.domen.NicknameGenerator
 import com.netchar.nicknamer.domen.models.Nickname
 
-interface NicknameGeneratorService {
-    /**
-     * Generates random nickname based on [config]
-     */
-    suspend fun generateNickname(config: Config) : Nickname
-
+interface NicknameGeneratorService : NicknameGenerator {
     fun addToFavorites(nickname: Nickname)
 
     fun removeFromFavorites(nickname: Nickname)
@@ -31,13 +27,4 @@ interface NicknameGeneratorService {
     fun getFavoriteNicknames(): List<Nickname>
 
     fun isFavorite(nickname: Nickname) : Boolean
-
-    data class Config(
-        val nicknameLength: Int = 5,
-        val gender: Gender = Gender.MALE,
-        val alphabet: Alphabet = Alphabet.LATIN
-    ) {
-        enum class Gender(val value: String) { MALE("male"), FEMALE("female") }
-        enum class Alphabet(val value: String) { LATIN("latin"), CYRILLIC("cyrillic") }
-    }
 }
