+++ app/src/main/java/com/netchar/nicknamer/domen/service/NicknameGeneratorServiceImpl.kt
@@ -16,6 +16,8 @@
 
 package com.netchar.nicknamer.domen.service
 
+import com.netchar.nicknamer.domen.NicknameGenerator
+import com.netchar.nicknamer.domen.NicknameGenerator.Companion.buildQueryKey
 import com.netchar.nicknamer.domen.NicknameRepository
 import com.netchar.nicknamer.domen.models.Nickname
 import com.netchar.nicknamer.domen.models.Nickname.Companion.orEmpty
@@ -23,12 +25,8 @@ import com.netchar.nicknamer.domen.models.Nickname.Companion.orEmpty
 class NicknameGeneratorServiceImpl(
         private val repository: NicknameRepository
 ) : NicknameGeneratorService {
-    companion object {
-        private const val PEOPLE_PREFIX = "people"
-    }
-
-    override suspend fun generateNickname(config: NicknameGeneratorService.Config): Nickname {
-        val key = "${PEOPLE_PREFIX}_${config.gender.value}_${config.alphabet.value}"
+    override suspend fun generateNickname(config: NicknameGenerator.Config): Nickname {
+        val key = config.buildQueryKey()
         val models = repository.getModels()
         return models[key]?.generateNickname(config.nicknameLength).orEmpty()
     }
@@ -49,3 +47,4 @@ class NicknameGeneratorServiceImpl(
         return repository.isFavorite(nickname)
     }
 }
+
